+++ docs/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 
 ## [Unreleased]
 
+### Added
+
+- Added a way to echo the commands to be executed
+
+### Removed
+
+- Dropped python 3.8 support
+
 ## [6.1.0] - 2024-04-02
 
 ### Added